Done with Unit Testing SQl Mock, adding gateway , few rpcs need to be added based on frontend needs

+++ server/admin.go
@@ -2,13 +2,32 @@ package server
 
 import (
 	"context"
+	"fmt"
 
-	"example.com/pet-project/proto"
+	"example.com/pet-project/gen/proto"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
+func (m *MoviesuggestionsServiceserver) AdminLogin(ctx context.Context, req *proto.AdminLoginRequest) (*proto.AdminLoginResponse, error) {
+	if req.Email == "" || req.Password == "" {
+		return nil, status.Errorf(codes.FailedPrecondition, "Please enter both email and password")
+	}
+	id, err := m.Db.AdminLogin(req)
+	if err != nil {
+		return nil, err
+	}
+	resp := &proto.AdminLoginResponse{
+		AdminId: id,
+	}
+	return resp, nil
+}
+
 func (m *MoviesuggestionsServiceserver) AddMovieToDatabase(ctx context.Context, req *proto.AddMovieToDatabaseRequest) (*proto.AddMovieToDatabaseResponse, error) {
+
+	if req.Name == "" || req.Imageurl == "" || req.Director == "" || req.Rating == 0 || req.Description == "" || req.ReleaseDate == "" || req.Movieott == "" || req.AdminId == 0 || req.Category == "" {
+		return nil, status.Errorf(codes.FailedPrecondition, "Please enter all the fields")
+	}
 
 	movie, err := m.Db.AddMovieToDatabase(req)
 	if err != nil {
@@ -34,6 +53,10 @@ func (m *MoviesuggestionsServiceserver) AddMovieToDatabase(ctx context.Context, 
 
 func (m *MoviesuggestionsServiceserver) DeleteMovieFromDatabase(ctx context.Context, req *proto.DeleteMovieFromDatabaseRequest) (*proto.DeleteMovieFromDatabaseResponse, error) {
 
+	if req.MovieId == 0 {
+		return nil, status.Errorf(codes.FailedPrecondition, "Please Enter Id of the movie to be deleted")
+	}
+
 	status, err := m.Db.DeleteMovieFromDatabase(req)
 	if err != nil {
 		resp := &proto.DeleteMovieFromDatabaseResponse{
@@ -51,11 +74,16 @@ func (m *MoviesuggestionsServiceserver) DeleteMovieFromDatabase(ctx context.Cont
 }
 
 func (m *MoviesuggestionsServiceserver) GetFeedBack(req *proto.GetFeedBackRequest, stream proto.MovieSuggestionsService_GetFeedBackServer) error {
-
+	fmt.Println("I am first")
+	if req.AdminId == 0 {
+		return status.Errorf(codes.FailedPrecondition, "Please enter adminId")
+	}
+	fmt.Println("I am debugging this!!")
 	feedbacks, err := m.Db.GetFeedBack(req)
 	if err != nil {
 		return err
 	}
+	fmt.Println("I am debugging after db getfeedback call!!")
 
 	for _, feedback := range feedbacks {
 		if err := stream.Send(&proto.GetFeedBackResponse{Description: feedback}); err != nil {
@@ -63,4 +91,25 @@ func (m *MoviesuggestionsServiceserver) GetFeedBack(req *proto.GetFeedBackReques
 		}
 	}
 	return nil
+}
+
+func (m *MoviesuggestionsServiceserver) GetFeeedBack(ctx context.Context,req *proto.GetFeeedBackRequest) (*proto.GetFeeedBackResponse,error){
+	if req.AdminId==0{
+		return nil,status.Errorf(codes.FailedPrecondition,"Please enter AdminId")
+	}
+
+	var feedBacks []string
+	feedBacks,err := m.Db.GetFeeedBack(req)
+	if err!=nil{
+		return nil,err
+	}
+
+	resp := &proto.GetFeeedBackResponse{}
+
+	for _,descp := range feedBacks{
+		f := descp
+		resp.Description = append(resp.Description,f)
+	}
+
+	return resp,nil
 }
