+++ server/feedbacks.go
@@ -3,12 +3,18 @@ package server
 import (
 	"context"
 
-	"example.com/pet-project/proto"
+	"example.com/pet-project/gen/proto"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 func (m *MoviesuggestionsServiceserver) GiveFeedBack(ctx context.Context, req *proto.GiveFeedBackRequest) (*proto.GiveFeedBackResponse, error) {
+	if req.UserId==0  || req.Description==""{
+		return nil,status.Errorf(codes.FailedPrecondition,"Enter both userId and description of the feedback")
+	}
 
 	feedback,err := m.Db.GiveFeedBack(req)
+	
 	if err!=nil{
 		return nil,err
 	}
@@ -25,6 +31,10 @@ func (m *MoviesuggestionsServiceserver) GiveFeedBack(ctx context.Context, req *p
 }
 
 func (m *MoviesuggestionsServiceserver) UpdateFeedBack(ctx context.Context, req *proto.UpdateFeedBackRequest) (*proto.UpdateFeedBackResponse, error) {
+
+	if req.UserId==0 || req.FeedbackId==0 || req.Description==""{
+		return nil,status.Errorf(codes.FailedPrecondition,"PLease enter all the fields")
+	}
 
 	feedback,err := m.Db.UpdateFeedBack(req)
 	if err!=nil{
@@ -43,6 +53,10 @@ func (m *MoviesuggestionsServiceserver) UpdateFeedBack(ctx context.Context, req 
 }
 
 func (m *MoviesuggestionsServiceserver) DeleteFeedBack(ctx context.Context, req *proto.DeleteFeedBackRequest) (*proto.DeleteFeedBackResponse, error) {
+
+	if req.UserId==0 || req.FeedbackId==0{
+		return nil,status.Errorf(codes.FailedPrecondition,"Please enter both userId and feedbackId")
+	}
 
 	status,err := m.Db.DeleteFeedBack(req)
 	if err!=nil{
