+++ server/likes.go
@@ -1,12 +1,18 @@
 package server
+
 import (
 	"context"
 
-	"example.com/pet-project/proto"
+	"example.com/pet-project/gen/proto"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 
 func (m *MoviesuggestionsServiceserver) AddMovieToLikes(ctx context.Context, req *proto.AddMovieToLikesRequest) (*proto.AddMovieToLikesResponse, error) {
+	if req.MovieId==0 || req.UserId==0{
+		return nil,status.Errorf(codes.FailedPrecondition,"Enter bith UserId and MovieId")
+	}
 
 	likes,err := m.Db.AddMovieToLikes(req)
 	if err!=nil{
@@ -25,6 +31,9 @@ func (m *MoviesuggestionsServiceserver) AddMovieToLikes(ctx context.Context, req
 }
 
 func (m *MoviesuggestionsServiceserver) RemoveMovieFromLikes(ctx context.Context, req *proto.RemoveMovieFromLikesRequest) (*proto.RemoveMovieFromLikesResponse, error) {
+	if req.UserId==0 || req.MovieId==0{
+		return nil,status.Errorf(codes.FailedPrecondition,"Enter UserId and MovieId")
+	}
 
 	status,err := m.Db.RemoveMovieFromLikes(req)
 	if err!=nil{
