+++ server/reviews.go
@@ -3,12 +3,15 @@ package server
 import (
 	"context"
 
-	"example.com/pet-project/proto"
+	"example.com/pet-project/gen/proto"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
 func (m *MoviesuggestionsServiceserver) AddReviewForMovie(ctx context.Context, req *proto.AddReviewRequest) (*proto.AddReviewResponse, error) {
+	if req.UserId==0 || req.MovieId==0 || req.Description==""{
+		return nil,status.Errorf(codes.FailedPrecondition,"Please enter UserId, MovieId and Description")
+	}
 
 	if req.Stars > 5 || req.Stars<=0{
 		return nil, status.Errorf(codes.FailedPrecondition, "Stars should be given in the range of 1 to 5")
@@ -34,6 +37,10 @@ func (m *MoviesuggestionsServiceserver) AddReviewForMovie(ctx context.Context, r
 
 func (m *MoviesuggestionsServiceserver) UpdateReviewForMovie(ctx context.Context, req *proto.UpdateReviewRequest) (*proto.UpdateReviewResponse, error) {
 
+	if req.UserId==0 || req.MovieId==0 || req.Description==""{
+		return nil,status.Errorf(codes.FailedPrecondition,"Enter userId movieId and Description to update")
+	}
+
 	review,err := m.Db.UpdateReviewForMovie(req)
 	if err!=nil{
 		return nil,err
@@ -52,6 +59,9 @@ func (m *MoviesuggestionsServiceserver) UpdateReviewForMovie(ctx context.Context
 }
 
 func (m *MoviesuggestionsServiceserver) DeleteReviewForMovie(ctx context.Context, req *proto.DeleteReviewRequest) (*proto.DeleteReviewResponse, error) {
+	if req.UserId==0 || req.MovieId==0{
+		return nil,status.Errorf(codes.FailedPrecondition,"Please enter valid UserId and MovieId")
+	}
 
 	status,err := m.Db.DeleteReviewForMovie(req)
 	if err!=nil{
